+++ nodejs/src/content/mod.rs
@@ -1,2 +1,3 @@
 pub mod context;
+pub mod requests;
 pub mod types;